backend/requirements.txt
fastapi
uvicorn
requests
transformers==4.37.2
numpy<2
torch==2.2.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
huggingface_hub==0.19.4
pydantic
cloudinary
pymongo
python-dotenv
kafka-python
python-multipart
